added vectorized NumPy energy routines
Numba is messing things up. I have remade the energy routines so that they are vectorized NumPy algorithms. They are not as fast - about half the speed. But, the results are more dependable.

This has still not quite matched the KolafaNezbeda equation of state, so there is a bug inside it somewhere.